Migrate App root component to TypeScript + hooks
